+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * fixed get_own_cpu_usage() to work on Windows
 * added __len__ to Optional
 * major bugfix in Optional
+* deprecated call_if_nnone